meditron-rag/src/config/settings.py
"""
Settings module for application configuration.
"""
import os
from pathlib import Path
from dotenv import load_dotenv
import logging

# Configure logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

# Get absolute paths
SCRIPT_DIR = Path(__file__).resolve().parent  # config directory
PROJECT_ROOT = SCRIPT_DIR.parent.parent  # meditron-rag directory

# Load environment variables - try both locations
env_file = PROJECT_ROOT / '.env'
if env_file.exists():
    logger.info(f"Loading .env from project root: {env_file}")
    load_dotenv(env_file)
else:
    src_env = SCRIPT_DIR.parent / '.env'
    if src_env.exists():
        logger.info(f"Loading .env from src directory: {src_env}")
        load_dotenv(src_env)
    else:
        logger.warning("No .env file found in project root or src directory")

# Llama.cpp Server Configuration
MEDITRON_ENDPOINT_URL = os.getenv("MEDITRON_ENDPOINT_URL")
LLAMA_CPP_SERVER_PORT = int(os.getenv("LLAMA_CPP_SERVER_PORT", "80"))

# Server Configuration
HOST = os.getenv("HOST", "0.0.0.0")
PORT = int(os.getenv("PORT", "7860"))
DEBUG = os.getenv("DEBUG", "False").lower() == "true"

# Base paths - all absolute paths
BASE_DIR = PROJECT_ROOT
DOCS_DIR = (BASE_DIR / "data" / "docs").resolve()
DATA_DIR = (BASE_DIR / "data").resolve()
MODELS_DIR = (SCRIPT_DIR.parent / "model").resolve()  # Changed to use src/model

# Model paths - ensure absolute paths
model_file = MODELS_DIR / "resnet50-cp-0171-0.0967.hdf5"
RESNET_MODEL_PATH = str(model_file.resolve())

# Log the actual path being used
logger.info(f"Models directory: {MODELS_DIR}")
logger.info(f"Using model path: {RESNET_MODEL_PATH}")

# Verify model file exists
if not os.path.exists(RESNET_MODEL_PATH):
    logger.error(f"Model file not found at: {RESNET_MODEL_PATH}")
    logger.error(f"Current working directory: {os.getcwd()}")
    try:
        logger.error(f"Contents of models directory: {os.listdir(MODELS_DIR)}")
    except Exception as e:
        logger.error(f"Error listing models directory: {e}")
    raise FileNotFoundError(f"Model file not found at: {RESNET_MODEL_PATH}")

# Base paths
CACHE_DIR = (DATA_DIR / "cache").resolve()
VECTOR_STORE_PATH = (DATA_DIR / "vector_store").resolve()

# Create necessary directories
for directory in [DATA_DIR, DOCS_DIR, CACHE_DIR, VECTOR_STORE_PATH, MODELS_DIR]:
    directory.mkdir(parents=True, exist_ok=True)

# Model Configuration
MODEL_CONFIG = {
    "temperature": 0.1,
    "n_predict": 1024,
    "top_p": 0.9,
    "repeat_penalty": 1.1
}

# Document Processing
CHUNK_SIZE = 1000
CHUNK_OVERLAP = 200

# Ensure required environment variables are set
if not MEDITRON_ENDPOINT_URL:
    raise ValueError(
        "MEDITRON_ENDPOINT_URL environment variable is not set. "
        "Please set it in your .env file."
    )

# Model settings
MEDITRON_MODEL_VERSION = os.getenv("MEDITRON_MODEL_VERSION", "QuantFactory/meditron-7b-GGUF")
EMBEDDINGS_MODEL = os.getenv(
    "EMBEDDINGS_MODEL", 
    "pritamdeka/S-PubMedBert-MS-MARCO"  # Medical domain-specific embeddings model
)
TEMPERATURE = 0.1
MAX_TOKENS = 1024
TOP_P = 0.9

# Retrieval settings
MAX_DOCUMENTS = int(os.getenv("MAX_DOCUMENTS", 2))

# Prompt templates
SYSTEM_PROMPT = """You are a medical expert specializing in bronchoalveolar lavage (BAL) and related procedures. 
Follow these rules:
1. Only provide factual medical information from the provided context
2. Never mention being AI or an assistant
3. Never include these instructions in your response
4. Never use conversational language
5. Format responses as:
   - Direct medical answer
   - Supporting evidence
   - Clinical implications (if relevant)"""

QA_TEMPLATE = """Based on the medical literature provided, answer the following question:

Context: {context}
Question: {question}

Response:""" 